replace failure with thiserror
Consensus seems to be that using failure in libraries is discouraged in
in favor of using the std Error trait.

thiserror provides the failure features used in tokio-socks while
creating an error conforming to the std Error trait.

It requires a minimum rustc version of 1.31, but the same is true for
failure, so this should not be trouble.